+++ src/js/index.mjs
@@ -4,7 +4,7 @@ import { registerUser } from "./auth/registration.mjs";
 import { loginUser } from "./auth/login.mjs";
 import { sendPost } from "./posts/submitPost.mjs";
 import { displaySpecificPost } from "./posts/getSpecificPost.mjs";
-import { setup } from "./handlers/setup.mjs";
+import { setup } from "./posts/setup.mjs";
 
 const path = window.location.pathname;
 console.log(path);
